plumbing: pktline, remove scanner/encoder
